requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
